requirements.txt
asgiref==3.10.0
atomicwrites==1.4.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.7
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.3.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
py==1.11.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
toml==0.10.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0

